pyproject.toml
[tool.poetry]
name = "TBD"
version = "0.1.0"
description = ""
authors = ["alivelimb"]

[tool.poetry.dependencies]
python = "^3.9"
functions-framework = "^3.2.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
mypy = "^0.982"
black = "^22.10.0"
isort = "^5.10.1"
pytest = "^7.1.3"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
pyproject-flake8 = "^5.0.4.post1"
types-python-dateutil = "^2.8.19"

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = [
  ".venv",
  "tests/"
]

[tool.flake8]
max-line-length = 120
ignore = "E203,W503,W504,"
exclude = [
  ".venv",
  "tests/",
]

[tool.black]
line-length = 120

[tool.isort]
line_length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
